Add Battery trackers
Add sensors to track the battery levels reported from Google Maps
Location Sharing. This includes a sensor for the level and a binary
sensor to indicate charging status, which is reflected in the icon for
the first sensor to be able to quickly see both values from the UI. This
is the first use of a package to implement a set of related entries.