Add googlehome component, adjust Info tab on UI
Add the google home component to track timers and alarms. Add this to
the Info tab and a custom card to the Home tab to only display active
timers/alarms. Adjust the Info tab to use a custom layout to fix some issues
with how it displays on different sized devices.